Move fetchL1L2LatestTransactions to fetch-data.ts
